Beginning new branch to convert pipeline to floating point.  So far, all Lab processing is done in floating point, and conversion from RGB to Lab is done via LUT and linear interpolation.

+++ rtengine/bilateral2.h
@@ -32,8 +32,8 @@
 
 // This seems ugly, but way faster than any other solutions I tried
 
-#define ELEM(a,b) (src[i - a][j - b] * ec[src[i - a][j - b]-src[i][j]+0x10000])
-#define SULY(a,b) (ec[src[i - a][j - b]-src[i][j]+0x10000])
+#define ELEM(a,b) (src[i - a][j - b] * ec[((int)(src[i - a][j - b]-src[i][j]+0x10000))])
+#define SULY(a,b) (ec[((int)(src[i - a][j - b]-src[i][j]+0x10000))])
 
 #define BL_BEGIN(a,b)   double scale = (a); \
                         int* ec = new int [0x20000]; \
